+++ android/src/main/kotlin/com/cheebeez/radio_player/service/MediaSessionCallback.kt
@@ -36,8 +36,7 @@ class MediaSessionCallback(private val radioPlayerService: RadioPlayerService) :
             val customCommands = mutableListOf<SessionCommand>()
             customCommands.add(SessionCommand(RadioPlayerService.CUSTOM_COMMAND_SET_STATION, Bundle.EMPTY))
             customCommands.add(SessionCommand(RadioPlayerService.CUSTOM_COMMAND_SET_CUSTOM_METADATA, Bundle.EMPTY))
-            customCommands.add(SessionCommand(RadioPlayerService.CUSTOM_COMMAND_SET_ITUNES_ARTWORK_PARSING, Bundle.EMPTY))
-            customCommands.add(SessionCommand(RadioPlayerService.CUSTOM_COMMAND_SET_IGNORE_ICY, Bundle.EMPTY))
+            customCommands.add(SessionCommand(RadioPlayerService.CUSTOM_COMMAND_RESET, Bundle.EMPTY))
 
             // Build the full set of available session commands (default + custom).
             val availableSessionCommandsBuilder = ConnectionResult.DEFAULT_SESSION_COMMANDS.buildUpon()
@@ -69,8 +68,10 @@ class MediaSessionCallback(private val radioPlayerService: RadioPlayerService) :
                 val title = args.getString("title")!!
                 val url = args.getString("url")!!
                 val imageData = args.getByteArray("image_data")
+                val parseStreamMetadata = args.getBoolean("parseStreamMetadata")!!
+                val lookupOnlineArtwork = args.getBoolean("lookupOnlineArtwork")!!
 
-                radioPlayerService.setStation(title, url, imageData)
+                radioPlayerService.setStation(title, url, imageData, parseStreamMetadata, lookupOnlineArtwork)
                 return Futures.immediateFuture(SessionResult(SessionResult.RESULT_SUCCESS))
             }
 
@@ -78,20 +79,13 @@ class MediaSessionCallback(private val radioPlayerService: RadioPlayerService) :
                 val artist = args.getString("artist")
                 val title = args.getString("title")
                 val artworkUrl = args.getString("artworkUrl")
-                
-                radioPlayerService.setMetadata(artist ?: "", title ?: "", artworkUrl ?: "")
+
+                radioPlayerService.setMetadata(artist, title, artworkUrl)
                 return Futures.immediateFuture(SessionResult(SessionResult.RESULT_SUCCESS))
             }
 
-            RadioPlayerService.CUSTOM_COMMAND_SET_ITUNES_ARTWORK_PARSING -> {
-                val enable = args.getBoolean("enable")
-                radioPlayerService.itunesArtworkParser = enable
-                return Futures.immediateFuture(SessionResult(SessionResult.RESULT_SUCCESS))
-            }
-
-            RadioPlayerService.CUSTOM_COMMAND_SET_IGNORE_ICY -> {
-                val ignoreIcy = args.getBoolean("ignore_icy")
-                radioPlayerService.ignoreIcy = ignoreIcy
+             RadioPlayerService.CUSTOM_COMMAND_RESET -> { 
+                radioPlayerService.reset()
                 return Futures.immediateFuture(SessionResult(SessionResult.RESULT_SUCCESS))
             }
         }