+++ ios/Classes/RadioPlayerService.swift
@@ -18,11 +18,11 @@ class RadioPlayerService: NSObject {
     var defaultArtwork: UIImage?
     private var metadataHash: String? = nil
 
-    weak var stateDelegate: RadioPlayerStateDelegate?
+    weak var playbackStateDelegate: RadioPlayerPlaybackStateDelegate?
     weak var metadataDelegate: RadioPlayerMetadataDelegate?
 
-    var ignoreIcy: Bool = false
-    var itunesArtworkParser: Bool = false
+    var parseStreamMetadata: Bool = true
+    var lookupOnlineArtwork: Bool = false
 
     /// Initialization
     override init() {
@@ -59,7 +59,8 @@ class RadioPlayerService: NSObject {
         }
 
         // Sets up observers for system notifications like audio interruptions.
-        NotificationCenter.default.addObserver(self, selector: #selector(handleInterruption), name: AVAudioSession.interruptionNotification, object: AVAudioSession.sharedInstance())
+        NotificationCenter.default.addObserver(self, selector: #selector(handleInterruption), name: AVAudioSession.interruptionNotification, 
+                object: AVAudioSession.sharedInstance())
     }
 
     deinit {
@@ -68,14 +69,26 @@ class RadioPlayerService: NSObject {
         UIApplication.shared.endReceivingRemoteControlEvents()
     }
 
-     /// Sets the media item for playback.
-    func setMediaItem(title: String, url: String, imageData: Data?) {
+    /// Sets the media item for playback.
+    func setStation(title: String, url: String, imageData: Data?, parseStreamMetadata: Bool, lookupOnlineArtwork: Bool) {
         streamTitle = title
         streamUrl = url
         defaultArtwork = imageData != nil ? UIImage(data: imageData!) : nil
 
-        MPNowPlayingInfoCenter.default().nowPlayingInfo = [MPMediaItemPropertyTitle: streamTitle, ]
+        // Update properties based on new parameters.
+        self.parseStreamMetadata = parseStreamMetadata
+        self.lookupOnlineArtwork = lookupOnlineArtwork
 
+        // Update Now Playing Info with station title initially.
+        var nowPlayingInfo: [String: Any] = [
+            MPMediaItemPropertyTitle: streamTitle,
+        ]
+        if let art = defaultArtwork {
+            nowPlayingInfo[MPMediaItemPropertyArtwork] = MPMediaItemArtwork(boundsSize: art.size, requestHandler: { _ in art })
+        }
+        MPNowPlayingInfoCenter.default().nowPlayingInfo = nowPlayingInfo
+
+        // Create and set the new player item.
         let playerItem = AVPlayerItem(url: URL(string: streamUrl)!)
         player.replaceCurrentItem(with: playerItem)
 
@@ -86,30 +99,31 @@ class RadioPlayerService: NSObject {
     }
 
     /// Updates the player's metadata with new track information.
-    func setMetadata(artist: String, songTitle: String, artworkUrl: String?) {
-        let trimmedArtist = artist.trimmingCharacters(in: .whitespacesAndNewlines)
-        let trimmedSongTitle = songTitle.trimmingCharacters(in: .whitespacesAndNewlines)
-        var trimmedArtworkUrl = artworkUrl?.trimmingCharacters(in: .whitespacesAndNewlines)
+    func setMetadata(artist: String?, songTitle: String?, artworkUrl: String?) {
+        let currentArtist = artist.nilIfEmpty()
+        let currentSongTitle = songTitle.nilIfEmpty()
+        var currentArtworkUrl = artworkUrl.nilIfEmpty()
 
         // Check if metadata has actually changed.
-        let currentMetadataHash = trimmedArtist + trimmedSongTitle
-        if currentMetadataHash == metadataHash { return }
-        metadataHash = currentMetadataHash
+        let newMetadataHash = (currentArtist ?? "_nil_") + (currentSongTitle ?? "_nil_")
+        if newMetadataHash == metadataHash { return }
+        metadataHash = newMetadataHash
 
         Task {
             // Parse artwork url from iTunes.
-            if (itunesArtworkParser && (artworkUrl ?? "").isEmpty) {
-                trimmedArtworkUrl = await ArtworkUtils.parseArtworkFromItunes(artist: trimmedArtist, track: trimmedSongTitle)
+            if (lookupOnlineArtwork && currentArtworkUrl == nil) {
+                currentArtworkUrl = await ArtworkUtils.parseArtworkFromItunes(
+                        artist: currentArtist ?? "", track: currentSongTitle ?? "")
             }
 
             // Download artwork.
-            var loadedArtwork = await ArtworkUtils.downloadImage(from: trimmedArtworkUrl)
+            var loadedArtwork = await ArtworkUtils.downloadImage(from: currentArtworkUrl)
             loadedArtwork = loadedArtwork ?? defaultArtwork
 
             // Update the now playing info.
             var nowPlayingInfo: [String: Any] = [
-                MPMediaItemPropertyArtist: trimmedArtist,
-                MPMediaItemPropertyTitle: trimmedSongTitle
+                MPMediaItemPropertyArtist: currentArtist,
+                MPMediaItemPropertyTitle: currentSongTitle
             ]
 
             if let image = loadedArtwork {
@@ -121,9 +135,9 @@ class RadioPlayerService: NSObject {
 
             // Send metadata to client.
             metadataDelegate?.radioPlayerDidUpdateMetadata(
-                artist: trimmedArtist ?? "",
-                title: trimmedSongTitle ?? "",
-                artworkUrl: trimmedArtworkUrl ?? "",
+                artist: currentArtist,
+                title: currentSongTitle,
+                artworkUrl: currentArtworkUrl,
                 artworkData: loadedArtwork?.jpegData(compressionQuality: 0.8)
             )
         }
@@ -133,7 +147,13 @@ class RadioPlayerService: NSObject {
     func play() {
          // If the player item is nil or has failed, try to set it up again.
         if player.currentItem == nil || player.currentItem?.isPlaybackBufferEmpty == true || player.currentItem?.status == .failed {
-             setMediaItem(title: streamTitle, url: streamUrl, imageData: defaultArtwork?.jpegData(compressionQuality: 1.0))
+            setStation(
+                title: streamTitle, 
+                url: streamUrl, 
+                imageData: defaultArtwork?.jpegData(compressionQuality: 1.0), 
+                parseStreamMetadata: self.parseStreamMetadata, 
+                lookupOnlineArtwork: self.lookupOnlineArtwork
+            )
         }
 
         player.play()
@@ -144,10 +164,16 @@ class RadioPlayerService: NSObject {
         player.pause()
     }
 
-    /// Stops playback.
-    func stop() {
+    /// Stops playback, removes notification, and fully resets player state for the next station.
+    func reset() {
         player.pause()
         player.replaceCurrentItem(with: nil)
+        MPNowPlayingInfoCenter.default().nowPlayingInfo = nil
+
+        streamTitle = nil
+        streamUrl = nil
+        defaultArtwork = nil
+        metadataHash = nil
     }
 
     /// Observes changes in player properties, like playback state.
@@ -158,20 +184,24 @@ class RadioPlayerService: NSObject {
         }
 
         let newStatus = AVPlayer.TimeControlStatus(rawValue: newStatusNumber.intValue)
-        var isPlaying = false
-        switch newStatus {
-            case .playing:
-                isPlaying = true
-            case .waitingToPlayAtSpecifiedRate:
-                isPlaying = true
-            case .paused:
-                isPlaying = false
-            case .none:
-                isPlaying = false
-            @unknown default:
-                isPlaying = false
+        var playbackState: String
+
+        if let status = newStatus {
+            switch status {
+                case .playing:
+                    playbackState = "playing"
+                case .waitingToPlayAtSpecifiedRate:
+                    playbackState = "buffering"
+                case .paused:
+                    playbackState = "paused"
+                @unknown default:
+                    playbackState = "unknown"
+            }
+        } else {
+            playbackState = "unknown" 
         }
-        stateDelegate?.radioPlayerDidChangeState(isPlaying: isPlaying)
+
+        playbackStateDelegate?.radioPlayerDidChangePlaybackState(playbackState: playbackState)
     }
 
     /// Handles audio session interruptions (e.g., phone calls).
@@ -204,19 +234,28 @@ extension RadioPlayerService: AVPlayerItemMetadataOutputPushDelegate {
 
     /// Processes incoming ICY (stream) metadata.
     func metadataOutput(_ output: AVPlayerItemMetadataOutput, didOutputTimedMetadataGroups groups: [AVTimedMetadataGroup], from track: AVPlayerItemTrack?) {
-        if ignoreIcy { return }
+        if !parseStreamMetadata { return }
 
-        let rawMetadata = groups.first.map({ $0.items })
+        guard let firstGroup = groups.first, !firstGroup.items.isEmpty else { return }
+        let metaItems = firstGroup.items
 
-        var artist = ""
-        guard var songTitle = rawMetadata?.first?.stringValue else { return }
-        let artworkUrlFromIcy = rawMetadata!.count > 1 ? rawMetadata![1].stringValue! : ""
+        guard var streamTitle: String = metaItems.first?.stringValue?.nilIfEmpty() else { return }
+
+        var artist: String? = nil
+        var songTitle: String? = nil
+        var artworkUrlFromIcy: String? = nil
+
+        if metaItems.count > 1 {
+            artworkUrlFromIcy = metaItems[1].stringValue?.nilIfEmpty()
+        }
 
         // Check if songTitle is in "Artist - Title" format
-        let parts = songTitle.components(separatedBy: " - ")
+        let parts = streamTitle.components(separatedBy: " - ")
         if parts.count == 2 {
-            artist = parts[0]
-            songTitle = parts[1]
+            artist = parts[0].nilIfEmpty()
+            songTitle = parts[1].nilIfEmpty()
+        } else {
+            songTitle = streamTitle.nilIfEmpty() // streamTitle is already nilIfEmpty checked
         }
 
         // Update metadata